auto-boxing 1.6 bug fixed
git-svn-id: svn+ssh://svn.msr-inria.inria.fr/var/lib/svn/repository/tla/trunk/tla2-inria@12824 76a6fc44-f60b-0410-a9a8-e67b0e8fc65c

+++ tlatools/src/tla2sany/parser/TLAplusParser.java
@@ -801,7 +801,7 @@ private int levelOfProofStepLexeme(Token tok){
   String im = tok.image ;
   if (im.substring(1,2).equals("*")) {return -1;} ;
   if (im.substring(1,2).equals("+")) {return -2;} ;
-  return new Integer(im.substring(1, im.indexOf('>'))) ;
+  return new Integer(im.substring(1, im.indexOf('>'))).intValue() ;
  }
 
 /***************************************************************************
